third-party/milton-1.4.2/milton-api/src/test/java/com/bradmcevoy/http/UtilsTest.java
package com.bradmcevoy.http;

import junit.framework.TestCase;

/**
 *
 * @author brad
 */
public class UtilsTest extends TestCase {
    
    public UtilsTest(String testName) {
        super(testName);
    }


    public void testPercentEncode() {
        for( int i=0; i<80; i++ ) {
            String s = String.valueOf((char)i);
            System.out.println(i + " = " + s);
        }
        assertEquals("", Utils.percentEncode(""));
        assertEquals("abc", Utils.percentEncode("abc"));
        assertEquals("%20", Utils.percentEncode(" "));
        assertEquals("ampersand%26", Utils.percentEncode("ampersand&"));
        assertEquals("0", Utils.percentEncode("0"));
        assertEquals("2009-01_02", Utils.percentEncode("2009-01_02"));

        // check decode simple cases
        assertEquals("abc", Utils.decodePath("abc"));
        assertEquals("/abc", Utils.decodePath("/abc"));

        // this string seems to encode differently on different platforms. this
        // isnt ideal and will hopefully be corrected, but in the mean time
        // its good enough if it 'round trips' Ie encode + decode = original
        String originalUnencoded = "neé";
        System.out.println("encoding: " + originalUnencoded);
        String encoded = Utils.percentEncode(originalUnencoded);
        System.out.println("encoded to: " + encoded);
        String decoded = Utils.decodePath(encoded);
        System.out.println("decoded to: " + decoded);
        assertEquals(originalUnencoded, decoded);
    }


    public void testDecodeHref() {
        String href = "/";
        String result = Utils.decodePath(href);
        assertEquals(result, href);

        href = "/with%20space";
        result = Utils.decodePath(href);
        assertEquals("/with space", result);

    }
}
